move more search-related code into search.go
